+++ mern-vb-frontend/jest.config.js
@@ -1,0 +1,15 @@
+// jest.config.js
+export default {
+  testEnvironment: 'jsdom',
+  verbose: true,
+  setupFilesAfterEnv: ['<rootDir>/src/setupTests.js'],
+  moduleNameMapper: {
+    '\\.(css|less|scss|sass)$': 'identity-obj-proxy',
+    '\\.(jpg|jpeg|png|gif|svg)$': '<rootDir>/__mocks__/fileMock.js'
+  },
+  testMatch: ['**/src/**/*.test.js', '**/?(*.)+(spec|test).js'],
+  coveragePathIgnorePatterns: ['/node_modules/', '/dist/'],
+  transform: {
+    '^.+\\.[jt]sx?$': 'babel-jest'
+  }
+};