+++ mern_vb_backend/jest.config.js
@@ -1,0 +1,7 @@
+// jest.config.js
+module.exports = {
+  testEnvironment: 'node',
+  verbose: true,
+  testMatch: ['**/tests/**/*.test.js', '**/?(*.)+(spec|test).js'],
+  coveragePathIgnorePatterns: ['/node_modules/', '/dist/'],
+};