+++ mern_vb_backend/models/BankBalance.js
@@ -3,5 +3,5 @@ const mongoose = require('mongoose');
 const bankBalanceSchema = new mongoose.Schema({
   balance: { type: Number, required: true, default: 0 }
 });
-
+ 
 module.exports = mongoose.model('BankBalance', bankBalanceSchema);